Move 'DatabaseAdapter' into a separate file
