+++ jardin/query.py
@@ -2,7 +2,7 @@ import argparse
 import os
 import inspect
 
-from jardin.database import DatabaseAdapter
+from jardin.database.database_adapter import DatabaseAdapter
 from jardin.database.datasources import Datasources
 from jardin.tools import stack_marker
